Add tests for bonus and refund
